requirements.txt
fastapi==0.100.0
uvicorn[standard]==0.23.2
selenium==4.21.0
webdriver-manager==4.0.2
python-dotenv==1.0.0        # para cargar variables de entorno
